Please Help!!! enemy code errors
I was making this code and wel...maybe I am not too good working with the NavMesh and the transform values...It doesnt have erros in unity. But it doesnt work. The enemy looks at the target (and changes the animation bools) but....she (the enemy) doesnt follow the target.
*The names are in spanish but...well i dont know if it will be a problem.